refactor(users, customers): added users filter functionality and rearranged users and customers
