+++ resources/js/pages/users/page.tsx
@@ -1,30 +1,18 @@
 import { router, usePage } from "@inertiajs/react";
-import { useState } from "react";
 
 import Layout from "@/layouts/private-layout";
 import { UserTable } from "./components/user-list";
-import { AddUserDialog } from "./components/user-dialog";
 import { UserLogsTable } from "./components/user-logs";
 
 import SectionHeader from "@/components/section-header";
 import SectionContent from "@/components/section-content";
-import DatePicker from "@/components/composables/date-picker";
-import Button from "@/components/composables/button";
-import Input from "@/components/composables/input";
-
-import {
-    Popover,
-    PopoverContent,
-    PopoverTrigger,
-} from "@/components/ui/popover";
 
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
-import { ListFilter } from "lucide-react";
-
 import type { User, Log } from "@/lib/interface/types";
-
-import { users } from "./data/users";
+import Filter from "./components/filter-popover";
+import { AddUserDialog } from "./components/user-dialog";
+import AlertNotification from "./components/alert-notification";
 
 type UsersProps = {
     users: User[];
@@ -32,12 +20,6 @@ type UsersProps = {
 };
 
 export default function Users({ users, logs }: UsersProps) {
-    const { props }: any = usePage();
-    const { flash } = props as any;
-
-    const [startDate, setStartDate] = useState<Date>();
-    const [endDate, setEndDate] = useState<Date>();
-
     const handleView = (user: User) => {
         alert(`Viewing ${user.name}`);
     };
@@ -52,82 +34,19 @@ export default function Users({ users, logs }: UsersProps) {
         <div>
             <main>
                 <Layout>
-                    {flash.success && (
-                        <div className="m-2 rounded-md p-3 bg-green-100 text-green-700 ">
-                            {flash.success}
-                        </div>
-                    )}
-                    {flash.delete && (
-                        <div className="m-2 rounded-md p-3 bg-red-100 text-red-700 ">
-                            {flash.delete}
-                        </div>
-                    )}
-
                     <Tabs defaultValue="users" className="w-full">
                         <SectionHeader className="flex items-center justify-between">
+                            <div className="fixed top-4 right-4 z-50 space-y-2 w-80">
+                                <AlertNotification />
+                            </div>
                             <TabsList>
                                 <TabsTrigger value="users"> Users</TabsTrigger>
                                 <TabsTrigger value="userLogs">
                                     User Logs
                                 </TabsTrigger>
                             </TabsList>
-
-                            <div className="flex items-center gap-4">
-                                <Popover>
-                                    <PopoverTrigger className="hover:bg-gray-100 p-2 cursor-pointer rounded border">
-                                        <ListFilter size={16} />
-                                    </PopoverTrigger>
-                                    <PopoverContent className="space-y-4 w-72">
-                                        <h1 className="font-semibold text-sm">
-                                            Filter Actions
-                                        </h1>
-                                        <Input
-                                            label="Code"
-                                            placeholder="Code"
-                                        />
-                                        <Input
-                                            label="Name"
-                                            placeholder="Name"
-                                        />
-                                        <Input
-                                            label="Municipal"
-                                            placeholder="Municipal"
-                                        />
-                                        <Input
-                                            label="Meter"
-                                            placeholder="Meter"
-                                        />
-                                        <Input
-                                            label="Brgy"
-                                            placeholder="Brgy"
-                                        />
-                                        <div className="grid grid-cols-2 gap-2">
-                                            <DatePicker
-                                                label="Start"
-                                                date={startDate}
-                                                setDate={setStartDate}
-                                                placeholder="Start date"
-                                            />
-                                            <DatePicker
-                                                label="End"
-                                                date={endDate}
-                                                setDate={setEndDate}
-                                                placeholder="End date"
-                                            />
-                                        </div>
-                                        <div className="mt-2 grid grid-cols-2 gap-2">
-                                            <Button
-                                                variant="outline"
-                                                shape="rounded"
-                                            >
-                                                Cancel
-                                            </Button>
-                                            <Button shape="rounded">
-                                                Submit
-                                            </Button>
-                                        </div>
-                                    </PopoverContent>
-                                </Popover>
+                            <div className="flex gap-6 items-center">
+                                <Filter />
                                 <AddUserDialog />
                             </div>
                         </SectionHeader>
